+++ tests/earl-report.js
@@ -11,7 +11,7 @@
  *
  * @param options {Object} reporter options
  *          id: {String} report id
- *          env: {String} environment description
+ *          env: {Object} environment description
  */
 function EarlReport(options) {
   let today = new Date();
@@ -24,6 +24,8 @@ function EarlReport(options) {
   this.now.setMilliseconds(0);
   this.id = options.id;
   this.env = options.env;
+  // test environment
+  this._environment = null;
   /* eslint-disable quote-props */
   this._report = {
     '@context': {
@@ -69,24 +71,20 @@ function EarlReport(options) {
       'foaf:homepage': 'https://digitalbazaar.com/'
     },
     'doap:release': {
-      'doap:name': '',
       'doap:revision': '',
       'doap:created': today
     },
     'subjectOf': []
   };
   /* eslint-enable quote-props */
-  this._report['@id'] +=
-    '#' + this.id;
-  this._report['doap:name'] +=
-    ' ' + this.id + (this.env ? ' ' + this.env : '');
-  this._report['dc:title'] +=
-    ' ' + this.id + (this.env ? ' ' + this.env : '');
+  if(this.env && this.env.version) {
+    this._report['doap:release']['doap:name'] = this.env.version;
+  }
 }
 
 EarlReport.prototype.addAssertion = function(test, pass, options) {
   options = options || {};
-  this._report.subjectOf.push({
+  const assertion = {
     '@type': 'earl:Assertion',
     'earl:assertedBy': this._report['doap:developer']['@id'],
     'earl:mode': 'earl:automatic',
@@ -94,10 +92,19 @@ EarlReport.prototype.addAssertion = function(test, pass, options) {
     'earl:result': {
       '@type': 'earl:TestResult',
       'dc:date': this.now.toISOString(),
-      'earl:outcome': pass ? 'earl:passed' : 'earl:failed',
-      ...options.extra
+      'earl:outcome': pass ? 'earl:passed' : 'earl:failed'
     }
-  });
+  };
+  if(options.benchmarkResult) {
+    const result = {
+      ...options.benchmarkResult
+    };
+    if(this._environment) {
+      result['jldb:environment'] = this._environment['@id'];
+    }
+    assertion['jldb:result'] = result;
+  }
+  this._report.subjectOf.push(assertion);
   return this;
 };
 
@@ -107,6 +114,83 @@ EarlReport.prototype.report = function() {
 
 EarlReport.prototype.reportJson = function() {
   return JSON.stringify(this._report, null, 2);
+};
+
+/* eslint-disable quote-props */
+const _benchmarkContext = {
+  'jldb': 'http://json-ld.org/benchmarks/vocab#',
+  'rdfs': 'http://www.w3.org/2000/01/rdf-schema#',
+  'xsd': 'http://www.w3.org/2001/XMLSchema#',
+
+  // environment description
+  'jldb:Environment': {'@type': '@id'},
+
+  // per environment
+  // architecture type
+  // ex: x86
+  'jldb:arch': {'@type': 'xsd:string'},
+  // cpu model description (may show multiple cpus)
+  // ex: 'Intel(R) Core(TM) i7-4790K CPU @ 4.00GHz'
+  'jldb:cpu': {'@type': 'xsd:string'},
+  // count of cpus, may not be uniform, just informative
+  'jldb:cpuCount': {'@type': 'xsd:integer'},
+  // platform name
+  // ex: linux
+  'jldb:platform': {'@type': 'xsd:string'},
+  // runtime name
+  // ex: Node.js, Chromium, Ruby
+  'jldb:runtime': {'@type': 'xsd:string'},
+  // runtime version
+  // ex: v14.19.0
+  'jldb:runtimeVersion': {'@type': 'xsd:string'},
+  // arbitrary comment
+  'jldb:comment': 'rdfs:comment',
+
+  // benchmark result
+  'jldb:BenchmarkResult': {'@type': '@id'},
+
+  // use in earl:Assertion, type jldb:BenchmarkResult
+  'jldb:result': {'@type': '@id'},
+
+  // per BenchmarkResult
+  'jldb:environment': {'@type': '@id'},
+  'jldb:hz': {'@type': 'xsd:float'},
+  'jldb:rme': {'@type': 'xsd:float'}
+};
+/* eslint-enable quote-props */
+
+// setup @context and environment to handle benchmark data
+EarlReport.prototype.setupForBenchmarks = function(options) {
+  // add context if needed
+  if(!Array.isArray(this._report['@context'])) {
+    this._report['@context'] = [this._report['@context']];
+  }
+  if(!this._report['@context'].some(c => c === _benchmarkContext)) {
+    this._report['@context'].push(_benchmarkContext);
+  }
+  if(options.testEnv) {
+    // add report environment
+    const fields = [
+      ['arch', 'jldb:arch'],
+      ['cpu', 'jldb:cpu'],
+      ['cpuCount', 'jldb:cpuCount'],
+      ['platform', 'jldb:platform'],
+      ['runtime', 'jldb:runtime'],
+      ['runtimeVersion', 'jldb:runtimeVersion'],
+      ['comment', 'jldb:comment']
+    ];
+    const _env = {
+      '@id': '_:environment:0'
+    };
+    for(const [field, property] of fields) {
+      if(options.testEnv[field]) {
+        _env[property] = options.testEnv[field];
+      }
+    }
+    this._environment = _env;
+    this._report['@included'] = this._report['@included'] || [];
+    this._report['@included'].push(_env);
+  }
 };
 
 module.exports = EarlReport;